requirements.txt
fastapi==0.83.0
uvicorn
SQLAlchemy
pymysql
httpx
python-dotenv
